Database

DataBase

ModMashup.Database — Type.

Store general and in-depth information for network integration and label propagation.

Fields

string_nets::Vector{String}: Similarity networks name.

disease::OneHotAnnotation: Disease annotation for patients.

n_patients::Int: The number of patients in the databse.

patients_index::Dict{String,Int}: Map patient name to their id.

inverse_index::Dict{Int,String}: Map patient id to their name.

num_cv::Int: The number of cross validation round. Default is 10.

query_attr::Int: Set the annotaion for query . Default is 1.

string_querys::Vector{String}: A list of query filename.

smooth::Int: Perform smooth in the simialarty or not. Default is true.

thread::Int: Number of thread for parallel computing. Default it 1.

Constructor

Database(network_dir, target_file, id, query_dir)

Create new Database.

Example

# get example data directory
example_data_dir = joinpath(Pkg.dir("ModMashup"), "test/data")

# dir should be a directory containing similairty networks flat file.
network_dir = joinpaht(example_data_dir,"networks")

# target_file should be a flat file contains disaese for patient
target_file = joinpaht(example_data_dir,"target.txt")

# Directory where a list of query flat files are located using the 
# same format and naming manner with genemania query.
query_dir = example_data_dir

# Id file contains all the name of patients.
id = joinpaht(example_data_dir,"ids.txt")

# Other setting
## Do smooth in the network or not for mashup integration.
smooth = true
## The number of thread choosen to perform computation.
thread = 2

# We are ready now, then just construct the database
database = ModMashup.Database(network_dir, target_file, 
                id, query_dir, smooth = smooth, 
                thread = thread)

source

ModMashup.SharedDatabase — Type.

SharedGeneMANIA

Shared genemania model for shared memory parallel computing (Still unfinished, left to do..)

source