Network Integration

Network integration

Integration model

The package provided two algorithm for network integration, one is MashupIntegration and another is GeneMANIAIntegration.

ModMashup.MashupIntegration — Type.

Modified Mashup algorithm for network integration. Inside MashupIntegration model, it contains all the result after mashup integration.

Fields

β::Vector: Beta vector as a result of linear regression.

H::Matrix: Rows of H represent patients embendding in networks.

net_weights::Vector: Normalized mean network weights

weights_mat::Matrix: Columns of weights_mat is computed network weights for each round of cross validation.

cv_query::Matrix: Columns of cv_query is query id for each round of cross validation.

singular_value_sqrt::Vector: singular value from mashup for dimensianal reduction.

tally::Vector{Int}: Network tally result

Constructor

MashupIntegration()

Create empty MashupIntegration model.

source

ModMashup.GeneMANIAIntegration — Type.

GeneMANIA lienar regression algorithm for network integration.

Fields

net_weights::Dict{String, Float64}: A dictionalry map network name to its final network weights result, which is same with GeneMANIA.jar.

normalized::Bool: Wether normlize the network weights

reg::Bool: Wether add regularization term to

source

Generic integration method

We have one generic function network_integration! to provide same interface for both mashup and genemania integration.

The database contains the input. After the computation, the result will be saved on the model.

ModMashup.network_integration! — Method.

network_integration!(model::MashupIntegration, database::GMANIA)

Implement modified mashup network integration. Result will be save in the model. See MashupIntegration for more information about the result.

Arguments

model::MashupIntegration: Label Propagation model.

database::Database: store general information about the patients.

Output

model::LabelPropagation: Result will be saved in the model fileds.

source

ModMashup.network_integration! — Method.

network_integration!(model::GeneMANIAIntegration, database::GMANIA)

Implement Raw mashup network integration. Result will be saved in the model. See GeneMANIAIntegration for more information about the result.

source