Util function

Util function

ModMashup.rwr — Function.

rwr(A::Matrix, restart_prob = 0.5)

Random walk with restart.

Arguments

A::Matrix: Initial matrix for random walk.

restart_prob: Restart probability. Default 0.5.

source

ModMashup.pca — Function.

pca(A::Matrix, num::Int64=size(A,2))

Perform PCA for a matrix.

Input: a matrix and needed

Output: pca vector and pca value

Example

julia> a = rand(5,5)
5×5 Array{Float64,2}:
 0.149599  0.318179  0.235567  0.779247  0.276985
 0.175398  0.109825  0.532561  0.723127  0.621328
 0.68087   0.639779  0.754652  0.781525  0.264776
 0.77962   0.446314  0.805693  0.88001   0.655808
 0.19243   0.43587   0.945708  0.109192  0.196602

julia> pca_value, pca_vector = pca(a)
([2.77556e-17,0.000524192,0.0396649,0.113626,0.128647],
[0.483429 0.397074 … -0.376334 -0.679859; -0.738733 0.15917 … -0.379275 -0.170866; … ;
 -0.0942359 -0.593824 … 0.444543 -0.643074; -0.457645 0.345674 … 0.1949 -0.306201])

source

ModMashup.build_index — Method.

build_index(index_file::String)

Get two dictionary, one map patients name to its id, another map patient id to its name.

Example

# get example data directory
example_data_dir = joinpath(Pkg.dir("ModMashup"), "test/data")

# Id file contains all the name of patients.
id = joinpaht(example_data_dir,"ids.txt")

# Build the index
patients_index, inverse_index = build_index(id)

source

ModMashup.parse_target — Method.

parse_target(target, patients_index)

Get a vector of annotation for patients. (+1 for interested, -1 for others)

Example

# get example data directory
example_data_dir = joinpath(Pkg.dir("ModMashup"), "test/data")

# Id file contains all the name of patients.
id = joinpaht(example_data_dir,"ids.txt")

# Build the patient index
patients_index, inverse_index = build_index(id)

# target_file should be a flat file contains disaese for patient
target_file = joinpaht(example_data_dir,"target.txt")

# Build the annotation for each patients
annotation = parse_target(readdlm(target_file), patients_index)

source

ModMashup.parse_query — Method.

parse_query(query_file, patients_index)

Get query patient id from the query file.

Example

# get example data directory
example_data_dir = joinpath(Pkg.dir("ModMashup"), "test/data")

# Id file contains all the name of patients.
id = joinpaht(example_data_dir,"ids.txt")

# Build the patient index
patients_index, inverse_index = build_index(id)

# Query file using the same format with genemania query
query = joinpaht(example_data_dir,"query.txt")

# Build the annotation for each patients
query = parse_query(query, patients_index)

source

ModMashup.load_net — Method.

load_net(filename::String,
              database::Database)

Load similairity network from a flat file. Format patient_name patient_name simialirty_score. Use databse to map patient_name to internal id.

source

ModMashup.searchdir — Method.

searchdir(path,key)

Input: directory we want to search and the keyword.

Output: a list of files whose name contains the keyword provided.

source