Network Integration

Network integration

Constructor

The package provided two algorithm for network integration, one is MashupIntegration and another is RawMashupIntegration.

GeneMANIA.MashupIntegration — Type.

MashupIntegration(β, H, net_weights, weights_mat, cv_query, singular_value_squared, tally)

Inside MashupIntegration model, it contains all the result after mashup integration.

β::Vector: Beta vector as a result of linear regression.

H::Matrix: rows of H represent patients embendding in networks.

net_weights::Vector:

weights_mat::Matrix:

cv_query::Matrix: columns of cv_query is query id for each round of cross validation.

singular_value_squared::Vector:

tally::Vector{Int}: Network tally result

source

Description

Example

References