Pipeline

PipeLine

ModMashup.fit! — Method.

fit!(database::Database,
         it_model::MashupIntegration,
         lp_model::LabelPropagation)

Pipeline the mashup integration and label propagation in one function.

Arguments

database::Database: Database for computation

it_model::MashupIntegration: GeneMANIAIntegration model contains result from integration

lp_model::LabelPropagation: LabelPropagation model contains result from label propgation

source

ModMashup.fit! — Method.

fit!(database::Database,
         it_model::GeneMANIAIntegration,
         lp_model::LabelPropagation)

Pipeline the genemania integration and label propagation in one function.

Arguments

database::Database: Database for computation

it_model::GeneMANIAIntegration: GeneMANIAIntegration model contains result from integration

lp_model::LabelPropagation: LabelPropagation model contains result from label propgation

source