Util function

Util function

ModMashup.rwr — Function.

rwr(A::Matrix, restart_prob = 0.5)

Random walk with restart.

Arguments

A::Matrix: Initial matrix for random walk.

restart_prob: Restart probability. Default 0.5.

Outputs

Q::Matrix: matrix after random walk.

source

ModMashup.pca — Function.

pca(A::Matrix, num::Int64=size(A,2))

Perform PCA for a matrix.

Arguments

A::Matrix: matrix for PCA

keywords

num::Int64=size(A,2): Number of diensions selected.

Output

pca vector and pca value

eigenvalue::Vector: as name suggested. eigenvector::Matrix: columns represent eigen vector.

Example

julia> a = rand(5,5)
5×5 Array{Float64,2}:
 0.149599  0.318179  0.235567  0.779247  0.276985
 0.175398  0.109825  0.532561  0.723127  0.621328
 0.68087   0.639779  0.754652  0.781525  0.264776
 0.77962   0.446314  0.805693  0.88001   0.655808
 0.19243   0.43587   0.945708  0.109192  0.196602

julia> pca_value, pca_vector = pca(a)
([2.77556e-17,0.000524192,0.0396649,0.113626,0.128647],
[0.483429 0.397074 … -0.376334 -0.679859; -0.738733 0.15917 … -0.379275 -0.170866; … ;
 -0.0942359 -0.593824 … 0.444543 -0.643074; -0.457645 0.345674 … 0.1949 -0.306201])

source

ModMashup.build_index — Method.

build_index(index_file::String)

Get two dictionary, one map patients name to its id, another map patient id to its name.

Arguments

index_file::String:

Outputs

patients_index::Dict{String, Int}: map patientd name to its internal id.

inverse_index::Dict{Int, String}: map patientd internal id to its name.

Example

# get example data directory
example_data_dir = joinpath(Pkg.dir("ModMashup"), "test/data")

# Id file contains all the name of patients.
id = joinpaht(example_data_dir,"ids.txt")

# Build the index
patients_index, inverse_index = build_index(id)

source

parse_target(target, patients_index)

parse_query(query_file, patients_index)

ModMashup.load_net — Method.

load_net(filename::String,
              database::Database)

Load similairity network from a flat file. Format patient_name patient_name simialirty_score. Use databse to map patient_name to internal id.

Inputs

filename::String: similairty network filename.

database::Database: store general information.

Outputs

A::Matrix: similairty network as a matrix.

source

searchdir(path,key)