Pipeline

PipeLine

ModMashup.fit! — Method.

fit!(int_model::MashupIntegration,
         lp_model::LabelPropagation,
         database::Database)

Pipeline the mashup integration and label propagation in one function.

Arguments

database::Database: Database for computation

it_model::MashupIntegration: GeneMANIAIntegration model contains result from integration

lp_model::LabelPropagation: LabelPropagation model contains result from label propgation

Outputs

it_model::MashupIntegration: outpus stored in model fileds.

lp_model::LabelPropagation: outpus stored in model fileds.

source

ModMashup.fit! — Method.

fit!(database::Database,
         it_model::GeneMANIAIntegration,
         lp_model::LabelPropagation)

Pipeline the genemania integration and label propagation in one function.

Arguments

database::Database: Database for computation

it_model::GeneMANIAIntegration: GeneMANIAIntegration model contains result from integration

lp_model::LabelPropagation: LabelPropagation model contains result from label propgation

source